morello-toolchain: refactor class and configs
The LLVM toolchain shall now be selected by setting the TOOLCHAIN variable within
the recipe or globally in .conf file. The idea is that one will be able to toggle
between GCC/LLVM in the future. There is no detailed spec on how this will exactly
work at this point in time.

Signed-off-by: Pawel Zalewski <pzalewski@thegoodpenguin.co.uk>