musl: refactor recipe
The LLVM compiler-rt recipes lived with the musl library, which does not
make sense, place them where they belog: in devtools. Move musl to
meta-morello-toolchain as it is required to build compiler-rt. Enable
musl to use the new toolchain.

Linux image and therefore fvp configs are affected as well as MORELLO_ARCH
variable belongs to morello-toolchain layer.

Signed-off-by: Pawel Zalewski <pzalewski@thegoodpenguin.co.uk>