classes: add rpath to the compiler flags and fix paths
Use rpath to aid the development of the user space recipes.

Hard code /usr instead of using $prefix and $libdir as
theese will be very likely overwritten becasue we use a
custom sysroot path for purecap. If we do not hardcode the paths
then we might end up with a recursion.

Signed-off-by: Pawel Zalewski <pzalewski@thegoodpenguin.co.uk>